django-project-template/requirements/base.txt
Django==1.9.1
django-jet==0.1.4
djangorestframework
markdown
django-filter
django-model-utils
